% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcalibration_analysis.R
\name{analyze_calibration_deep}
\alias{analyze_calibration_deep}
\title{Create comprehensive calibration comparison visualizations}
\usage{
analyze_calibration_deep(calib_results, output_dir = "results_transfer")
}
\arguments{
\item{calib_results}{Output from calibrate_model()}

\item{output_dir}{Directory to save plots}
}
\value{
List of ggplot objects
}
\description{
Create comprehensive calibration comparison visualizations
}
