% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb_download.R
\name{analyze_vdjdb_composition}
\alias{analyze_vdjdb_composition}
\title{Analyze VDJdb data composition}
\usage{
analyze_vdjdb_composition(vdjdb)
}
\arguments{
\item{vdjdb}{Full VDJdb tibble}
}
\value{
Summary tibble
}
\description{
Provides detailed breakdown of VDJdb by species, antigen source, etc.
Useful for planning training strategy.
}
