% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_encoding.R
\name{atchley_encode_sequence}
\alias{atchley_encode_sequence}
\title{Physicochemical encoding using Atchley factors}
\usage{
atchley_encode_sequence(sequence, max_length = 25)
}
\arguments{
\item{sequence}{Single amino acid sequence}

\item{max_length}{Maximum sequence length (for padding)}
}
\value{
Matrix of physicochemical features
}
\description{
Physicochemical encoding using Atchley factors
}
