% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{calc_topk_accuracy}
\alias{calc_topk_accuracy}
\title{Calculate top-k accuracy}
\usage{
calc_topk_accuracy(probabilities, true_labels, k = 5)
}
\arguments{
\item{probabilities}{Matrix of predicted probabilities (samples x classes)}

\item{true_labels}{Vector of true class indices}

\item{k}{Value of k for top-k accuracy}
}
\value{
Top-k accuracy value
}
\description{
Calculate top-k accuracy
}
