% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb_preprocessing.R
\name{clean_epitope_sequences}
\alias{clean_epitope_sequences}
\title{Clean epitope sequences}
\usage{
clean_epitope_sequences(sequences)
}
\arguments{
\item{sequences}{Character vector of epitope sequences}
}
\value{
Cleaned sequences
}
\description{
Clean epitope sequences
}
