% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb_preprocessing.R
\name{compare_weight_methods}
\alias{compare_weight_methods}
\title{Preview weight distributions for different methods}
\usage{
compare_weight_methods(scores = 0:3)
}
\arguments{
\item{scores}{Vector of scores (or just 0:3 for preview)}
}
\value{
Tibble comparing all methods
}
\description{
Useful for deciding which weighting scheme to use.
}
