% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{compute_epitope_similarities}
\alias{compute_epitope_similarities}
\title{Compute cosine similarities between query and epitope embeddings}
\usage{
compute_epitope_similarities(query_embeddings, epitope_embeddings)
}
\arguments{
\item{query_embeddings}{Matrix of query CDR3 embeddings (n_queries x embedding_dim)}

\item{epitope_embeddings}{Matrix of epitope embeddings (n_epitopes x embedding_dim)}
}
\value{
Matrix of similarities (n_queries x n_epitopes)
}
\description{
Compute cosine similarities between query and epitope embeddings
}
