% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_architecture.R
\name{create_tcr_epitope_model}
\alias{create_tcr_epitope_model}
\title{Create TCR-Epitope model}
\usage{
create_tcr_epitope_model(
  embedding_dim = 128,
  cdr3_max_len = 25,
  epitope_max_len = 30,
  dropout = 0.3
)
}
\arguments{
\item{embedding_dim}{Dimension of learned embeddings}

\item{cdr3_max_len}{Maximum CDR3 sequence length}

\item{epitope_max_len}{Maximum epitope sequence length}

\item{dropout}{Dropout rate}
}
\value{
PyTorch model object
}
\description{
Create TCR-Epitope model
}
