% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_setup_config.R
\name{first_time_setup}
\alias{first_time_setup}
\title{Perform first-time setup of conda environment and Python packages}
\usage{
first_time_setup(
  env_name = "tcr_epitope",
  python_version = "3.10",
  use_gpu = FALSE,
  force_reinstall = FALSE
)
}
\arguments{
\item{env_name}{Name of the conda environment to create}

\item{python_version}{Python version to use}

\item{use_gpu}{If TRUE, install CUDA-enabled PyTorch}

\item{force_reinstall}{If TRUE, reinstall packages even if present}
}
\value{
TRUE if successful
}
\description{
This function should only be run once. It will:
\enumerate{
\item Install miniconda if not present
\item Create the tcr_epitope conda environment
\item Install all required Python packages
}
}
