% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{freeze_encoder_layers}
\alias{freeze_encoder_layers}
\title{Freeze encoder layers for fine-tuning}
\usage{
freeze_encoder_layers(model)
}
\arguments{
\item{model}{TCREpitopeModel}
}
\description{
Freeze encoder layers for fine-tuning
}
