% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{generate_evaluation_report}
\alias{generate_evaluation_report}
\title{Generate comprehensive evaluation report}
\usage{
generate_evaluation_report(
  training_result,
  evaluation_results,
  output_dir = "results"
)
}
\arguments{
\item{training_result}{Result from train_tcr_epitope_model()}

\item{evaluation_results}{Result from evaluate_model()}

\item{output_dir}{Directory to save plots}
}
\value{
List of all plots
}
\description{
Generate comprehensive evaluation report
}
