% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_encoding.R
\name{get_atchley_factors}
\alias{get_atchley_factors}
\title{Amino acid physicochemical properties (Atchley factors)}
\usage{
get_atchley_factors()
}
\value{
Named list of Atchley factor vectors
}
\description{
Five principal components capturing:
\enumerate{
\item Polarity/hydrophobicity
\item Secondary structure propensity
\item Molecular volume
\item Codon diversity
\item Electrostatic charge
}
}
