% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_encoding.R
\name{get_blosum62_matrix}
\alias{get_blosum62_matrix}
\title{BLOSUM62-derived amino acid similarity encoding}
\usage{
get_blosum62_matrix()
}
\value{
20x20 similarity matrix
}
\description{
BLOSUM62-derived amino acid similarity encoding
}
