% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_query_embeddings_batched}
\alias{get_query_embeddings_batched}
\title{Get embeddings for query sequences in batches}
\usage{
get_query_embeddings_batched(trainer, query_indices, batch_size = 256)
}
\arguments{
\item{trainer}{Trained TCRTrainer object}

\item{query_indices}{Integer matrix of sequence indices}

\item{batch_size}{Batch size for processing}
}
\value{
Matrix of embeddings
}
\description{
Get embeddings for query sequences in batches
}
