% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{get_top_predictions}
\alias{get_top_predictions}
\title{Get top prediction for each query (convenience function)}
\usage{
get_top_predictions(predictions)
}
\arguments{
\item{predictions}{Full predictions tibble from predict_epitopes()}
}
\value{
Tibble with only top-1 prediction per query
}
\description{
Get top prediction for each query (convenience function)
}
