% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_setup_config.R
\name{initialize_session}
\alias{initialize_session}
\title{Initialise Python environment for current R session}
\usage{
initialize_session(env_name = "tcr_epitope")
}
\arguments{
\item{env_name}{Name of the conda environment}
}
\value{
TRUE if successful
}
\description{
This function MUST be called at the very start of each R session,
BEFORE loading reticulate or performing any Python operations.
}
