% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb_download.R
\name{list_vdjdb_releases}
\alias{list_vdjdb_releases}
\title{List available VDJdb releases}
\usage{
list_vdjdb_releases()
}
\value{
Data frame of available releases with dates
}
\description{
List available VDJdb releases
}
