% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{load_calibration}
\alias{load_calibration}
\title{Load calibration results from model package}
\usage{
load_calibration(model_dir)
}
\arguments{
\item{model_dir}{Directory containing saved model}
}
\value{
Calibration parameters
}
\description{
Load calibration results from model package
}
