% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_persistence.R
\name{load_epitope_reference}
\alias{load_epitope_reference}
\title{Load epitope reference database}
\usage{
load_epitope_reference(input_path = "models/epitope_reference.rds")
}
\arguments{
\item{input_path}{Path to the reference file}
}
\value{
Epitope reference list
}
\description{
Load epitope reference database
}
