% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{merge_invalid_sequences}
\alias{merge_invalid_sequences}
\title{Merge predictions with invalid sequences}
\usage{
merge_invalid_sequences(predictions, original_cdr3, top_k)
}
\arguments{
\item{predictions}{Predictions for valid sequences}

\item{original_cdr3}{Original CDR3 vector (including invalid)}

\item{top_k}{Number of prediction ranks}
}
\value{
Complete predictions tibble
}
\description{
Merge predictions with invalid sequences
}
