% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_encoding.R
\name{onehot_encode_sequence}
\alias{onehot_encode_sequence}
\title{One-hot encoding for amino acid sequences}
\usage{
onehot_encode_sequence(sequence, max_length = 25)
}
\arguments{
\item{sequence}{Single amino acid sequence}

\item{max_length}{Maximum sequence length (for padding)}
}
\value{
Matrix of one-hot encoded sequence
}
\description{
One-hot encoding for amino acid sequences
}
