% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{plot_confidence_analysis}
\alias{plot_confidence_analysis}
\title{Analyze prediction confidence}
\usage{
plot_confidence_analysis(evaluation_results)
}
\arguments{
\item{evaluation_results}{Results from evaluate_model()}
}
\value{
ggplot object
}
\description{
Analyze prediction confidence
}
