% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{plot_confidence_distribution}
\alias{plot_confidence_distribution}
\title{Plot confidence distribution before and after calibration}
\usage{
plot_confidence_distribution(calib_results)
}
\arguments{
\item{calib_results}{Output from calibrate_model()}
}
\value{
ggplot object
}
\description{
Plot confidence distribution before and after calibration
}
