% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot confusion matrix for top epitopes}
\usage{
plot_confusion_matrix(evaluation_results, top_n = 15, idx_to_epitope = NULL)
}
\arguments{
\item{evaluation_results}{Results from evaluate_model()}

\item{top_n}{Number of top epitopes to show}

\item{idx_to_epitope}{Mapping from index to epitope}
}
\value{
ggplot object
}
\description{
Plot confusion matrix for top epitopes
}
