% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{plot_training_history}
\alias{plot_training_history}
\title{Plot training history}
\usage{
plot_training_history(history)
}
\arguments{
\item{history}{Training history from trainer}
}
\value{
ggplot object
}
\description{
Plot training history
}
