% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{save_calibration}
\alias{save_calibration}
\title{Save calibration results to model package}
\usage{
save_calibration(model_dir, calib_results)
}
\arguments{
\item{model_dir}{Directory containing saved model}

\item{calib_results}{Output from calibrate_model()}
}
\description{
Save calibration results to model package
}
