% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_encoding.R
\name{sequences_to_indices}
\alias{sequences_to_indices}
\title{Convert sequences to integer indices (for embedding layers)}
\usage{
sequences_to_indices(sequences, max_length = 25L)
}
\arguments{
\item{sequences}{Character vector of sequences}

\item{max_length}{Maximum sequence length}
}
\value{
Integer matrix of indices
}
\description{
Convert sequences to integer indices (for embedding layers)
}
