% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{summarize_predictions_by_epitope}
\alias{summarize_predictions_by_epitope}
\title{Summarize predictions by epitope}
\usage{
summarize_predictions_by_epitope(predictions, rank_cutoff = 1)
}
\arguments{
\item{predictions}{Full predictions tibble}

\item{rank_cutoff}{Only consider predictions at or above this rank}
}
\value{
Summary tibble of epitope frequencies
}
\description{
Summarize predictions by epitope
}
